+++ catroid/src/main/java/org/catrobat/catroid/devices/mindstorms/ev3/LegoEV3.java
@@ -48,7 +48,7 @@ public interface LegoEV3 extends Mindstorms, BluetoothDevice {
 
 	void setLed(int ledStatus);
 
-	int getSensorValue(Sensors sensor);
+	float getSensorValue(Sensors sensor);
 
 	MindstormsSensor getSensor1();
 	MindstormsSensor getSensor2();
