+++ catroid/src/main/java/org/catrobat/catroid/devices/mindstorms/nxt/sensors/NXTSensor.java
@@ -84,7 +84,7 @@ public abstract class NXTSensor implements MindstormsSensor {
 	protected final MindstormsConnection connection;
 
 	protected boolean hasInit;
-	protected int lastValidValue = 0;
+	protected float lastValidValue = 0;
 
 	public static final String TAG = NXTSensor.class.getSimpleName();
 
@@ -180,7 +180,7 @@ public abstract class NXTSensor implements MindstormsSensor {
 	}
 
 	@Override
-	public int getLastSensorValue() {
+	public float getLastSensorValue() {
 		return lastValidValue;
 	}
